refactor(orb-live): A1 - extract wire types & protocol constants (VTID-02869)
Lifts pure wire-protocol shapes and Vertex model constants out of the
16k-line orb-live.ts into shared modules under services/gateway/src/orb/live/.
No behavior change — same declarations, same callers, same characterization
test snapshots match byte-for-byte.

What moved:

  services/gateway/src/orb/live/types.ts (NEW)
    - ClientContext (already exported; re-exported from new home)
    - LiveSessionStartRequest
    - TtsRequest
    - LiveStreamAudioChunk, LiveStreamVideoFrame, LiveStreamMessage

  services/gateway/src/orb/live/protocol.ts (NEW)
    - VERTEX_LIVE_MODEL  ('gemini-live-2.5-flash-native-audio')
    - VERTEX_TTS_MODEL   ('gemini-2.5-flash-tts')
    - AUDIO_OUT_RATE_HZ  (24000)
    - AUDIO_IN_RATE_HZ   (16000)

What did NOT move (deliberately):
  - GeminiLiveSession, WsClientSession, OrbLiveSession — these wrap
    runtime state (Maps, sockets, AbortControllers). They move in A8
    (session lifecycle extraction).
  - VERTEX_PROJECT_ID, VERTEX_LOCATION — env-derived; they move in A2
    (config/env extraction).
  - LIVE_API_VOICES, GEMINI_TTS_VOICES — large lookup tables; either
    move with the system instruction in A3 or stay until a dedicated
    voice-config module.

Verification:
  - cd services/gateway && npx jest test/orb/live/characterization/
    Result: 133 passed, 9 snapshots match (the 9 snapshots from A0.1/A0.2
    + the structural assertions from A0.3, all unchanged).
  - tsc --noEmit: zero new errors (only pre-existing 'sharp' module
    declaration error on origin/main, unrelated to this PR).

Part of the orb-live.ts refactor plan
(.claude/plans/for-an-intelligent-context-aware-fluttering-mango.md,
Track A, step A1). A2 (config/env) ships next.
